[feature] Replace max_ball_height with chipped flag in ball speed event
At the time where the ball speed event is triggered, the ball has
usually not yet reached it's maximum height. Thus, only a prediction
could be set, or the event would need to be delayed.
As this information is only for visualization purposes, it has been
replaced by a simple flag that indicates if the ball was chipped.