Fix deadlock or bad data read in CI connection when sending data burst
When reading from a TCP connection, a buffered reader
was created for each read.
As it turns out, it can happen that the buffer reads more than
needed and the next buffer is missing that data.
It instead reads the package size from the payload, which
results in random behavior (lock or incomplete packages).