Set default behavior for game events to accept_majority
If there is more than one autoRef connected, the default behavior "accept" will produce duplicated game events, which is an unexpected behavior for the regular user.
With majority, this will not happen. It will also work with one autoRef. Game events will just added to the proposed events first and applied immediately afterwards.