Avoid getting stuck due to missing ball placement command
It is currently possible, especially during dribble fouls, that the ball is still within the 1m radius where no ball placement is required. But if it is moving, it can get out of this radius. Currently, the check for ball placement is only done immediately after the STOP command. So the game is stuck in a situation, where ball placement is required, but no command is sent.